unsuedfiles/batch_api/requirements.txt
openai==1.10.0
python-dotenv==1.0.0
pyodbc==4.0.39
tiktoken==0.4.0
datetime==5.2
uuid==1.30
pytest==7.4.3
flake8==6.1.0
mypy==1.5.1
ruff==0.1.3